+++ bold-pagos-en-linea/bold-co.php
@@ -3,7 +3,7 @@
  * Plugin Name: Bold pagos en linea 
  * Plugin URI: https://developers.bold.co/pagos-en-linea/boton-de-pagos/plugins/wordpress
  * Description: Recibe pagos en tu tienda de forma segura con los métodos de pago más usados y con la mejor experiencia para tus clientes.
- * Version: 3.0.4
+ * Version: 3.0.5
  * Author: Bold
  * Author URI: http://www.bold.co/
  * Network: true
@@ -60,7 +60,7 @@ use BoldPagosEnLinea\BoldConstants;
 
 // Función para registrar y cargar el script de botón de pago
 function bold_co_custom_header_code(): void {
-    wp_register_script('woocommerce_bold_payment_button_js', 'https://checkout.bold.co/library/boldPaymentButton.js', [], '3.0.4', true);
+    wp_register_script('woocommerce_bold_payment_button_js', BoldConstants::URL_CHECKOUT.'/library/boldPaymentButton.js', [], '3.0.5', true);
     wp_enqueue_script('woocommerce_bold_payment_button_js');
 }
 
@@ -139,7 +139,7 @@ function bold_co_uninstall(){
                 delete_option($option_key);
             }
         } catch (\Throwable $th) {
-            error_log($th->getMessage(), 3, WP_CONTENT_DIR . '/bold_button_event_log.txt' );
+            BoldCommon::logEvent("Error: " . $th->getMessage() . " in file " . $th->getFile() . " line " . $th->getLine());
         }
     }
 }